Webbrowser and Fixed Tree Highlighting
Reorganized the code, storing everything in the "visualizer" folder that can then be called as needed. If not within a Jupyter Notebook, the visualizer launches a webbrowser to view the tree (within Notebook experience unchanged). Also, fixed an error with the tree highlighting that occurred when an edge had disjunct regions of the along the chromosome. Highlighted edges are now brought to the top of the SVG.