Using initial positions from graphviz
Added option to calculate node positions with graphviz before handing off to the D3 force simulation. Found what appears to be a bug in how D3 initializes the locations of nodes; it requires x and y to both be defined, and defining x and fy doesn't seem to cut it. This was super annoying and difficult to find, so will submit an issue once I confirm a bit more.