Fixed looping draw method
The draw method was inadvertently updating the D3ARG object which made subsequent draws of the same tree sequence incorrect. Better separated variable names; removed "scaled" from variable names and replaced it with "_XY" where X and Y are the lower and upper bound of the variable. This should help with confusion around what bounds the variable used for scaling.