Hideable controls and fixed background
Gave the visualizer a fixed width; specifically with Jupyter Notebooks, the output will be scrollable if it is too large. At the moment, I like giving users control over SVG size, rather than it being dynamic to the screen size (the latter would also require JS node positioning rather than Python #23). I also added a Hide controls button to get those out of the way when not in use. This dashboard is definitely still in prototype and stylized buttons would make it look a lot better.