fix reheat issue and font sizes
Previously if you reheated a visualization then copied to clipboard, the fx values of non-sample nodes was set to None (null). This caused an issue if you wanted to save that figure to JSON and load using draw_D3(). Now it deletes the attribute instead of setting to None.

In CSS, font-size needed "px" after number. Now should be consistent across browsers.