Add sequence length endpoints
Adding labels to each of the breakpoints gets cluttered but providing the endpoints seems like a good way of getting across the sequence length included in the ARG. I may add a popup when you highlight a breakpoint that specifies the bounds of that region. I still haven't looked into what to do when the regions are so small that they are pretty much invisible.