Dynamic vnub in edge
`vnub` the part of the edge that comes in from above a bifurcation node or below a recombination node. Previously, the `vnub` was fixed at 20px and if the nodes were too close together vertically, it would lead to an ugly overlap. Now `vnub` is set to either 20 or the half the vertical distance between the two nodes, whichever is smaller.