combining edges from the same child/parent combos
During tskit meeting, we talked about converting the ARG from a MultiDiGraph, where there can be multiple connections between the same parent and children, to a DiGraph. Previously, the each link (edge) in the visualization had a "left" and "right" attribute. This method has the restriction of only contiguous regions. Those two attributes have been replaced with a "bounds" attribute which is a string with the format "left-right left-right left-right...". This allows a link to be found in multiple regions. Link filtering during tree highlighting has been updated to account for this new attribute. There aren't any checks in the visualizer code to ensure that bounds make sense; this must be done within the tskit.TreeSequence. ts.breakpoints() is still used to set tree highlighting blocks.