#44 completed
Added arced edges for diamonds to better differentiate between the two lineages. This was particularly important for completing #43 and was first suggested in #42. I couldn't figure out how to get d3.path() to work properly, so this is currently coded directly by an SVG path string. I've always found those paths to be a bit confusing to read, so documentation may be needed in the event we want to change the amount of curve (currently hardcoded).