Fixing set_node_labels()
set_node_labels() was not working as intended. It previously used the index of the node in the D3ARG nodes list rather than the node's "id". This becomes an issue when there are recombination nodes. Also made node labels consistently strings, rather than having some as strings and others ints.