Changing how samples are ordered and positioned
This pertains to three issues (#48, #53, #55) relating to samples nodes (tips). I've added a "Space Samples" button to the dashboard that evenly distributes sample nodes along the x-axis after they've been moved. This cleans up the ARG nicely and is much easier/more accurate than manually positioning them. This functionality is also incorporated into the "Reheat Simulation" button, though having it separate from reheating seems helpful. Users can now specify the order of the sample nodes when drawing with the optional parameter `D3ARG.draw(..., sample_order=[...]). By default, it is the order of the first tree in the tree sequence. I've changed the backend of how sample nodes are positioned along the x-axis, but there aren't any visible changes for users. Lastly, you can now visualize the paths above a single sample and that sample is now positioned in the center of the x-axis.

All of this performs a bit more of the positioning in JS rather than Python, which I believe is best as it opens up opportunity for more user interactions. The visualizer is still heavily intertwined with tskit and node flags (even more so now with the changes) so this does limit its robustness with non-standard ARGs, particularly those that are not flagged properly.