Add three save methods (JSON, SVG, and PNG)
Relating to #59. New dashboard symbol for downloading. Copying to clipboard is gone now and has been replaced with the ability to save directly to a JSON file. SVG and PNG saving is now also possible and styles are preserved. I'm not entirely sold that the text buttons are clear enough as clickable choices, so may need to revisit this. This now requires two more JavaScript libraries (FileSaver.js and canvas-toBlob.js).