working minimal example for Pyodide
Put together a minimal example using Pyodide (not Pyscript). This does not have all of the functionality of the package, but still gives an idea of the potential. I also found out that the sample_order does not appear to be working properly. I need to check that in the package.