changing to pandas.DataFrames
This is a major change and is not backwards compatible! As we add more customization to the figures, I didn't like the ever growing number of optional parameters that could be passed to D3ARG.draw(). Previously, nodes, edges, and breakpoints were stored as lists of dictionaries. Now, they are pandas.DataFrames and include both info and styling. This now gives users control over the styling of each node individually (though this functionality still needs to be coded). The D3ARG.from_json() does not currently work (this is next on the TODO list). Advantages of this is that you have greater customization and styles can be preserved between figures. Disadvantages are the pandas dependency and larger JSON passed from Python to D3 as it stores styling for each node now. Once I finish with the above features, I'll send around for feedback.