better implementation of recomb positioning
The previous version worked if you only dragged the child node. Now you can drag either node. This permanently locks these nodes together; it may reduce some of the generality, but it is how I always style my ARGs so I haven't come up with an alternative use case. Feedback would be helpful here, will post on Mastodon. I've also gone a step farther and created more rules for locking together nodes. If a coalescent node has only one child, then it will be locked as well, unless that coalescent node is a root. This avoids multiple roots trying to be placed at the same point relative to a recombination node (common in chopped SLiM simulations). Pretty happy with this implementation. There is a slight delay when dragging the child node compared to the recombination node, but this doesn't bother me too much.