reworking y axis ticks
There was often a lot of overlap in tick labels when using `y_axis_scale="time"` or `y_axis_scale="log_time"` as the node timings were so close together. This now limits evenly spaces the tick marks for those scales, while keeping 0 and the max_time. This should avoid the issues mentioned in #8