node (subgraph) view
This function is based off of Figure 6 from the sc2ts paper. This function is very similar to the standard `draw()`, but you need to provide a node ID which will be the center of the subgraph. At the moment, it also doesn't have quite as many optional styling parameters; graphs are always drawn with tree_highlighting=False, y_axis_scale="rank" and edge_type="line", variable_edge_width=False, and include_underlink=True.