Google Colab compatibility
The running_in_notebook() function was not recognizing that Google Colabs was a notebook due to a different shell name. I have added an extra condition for this scenario. Because their are still relatively few environments, I don't think that it is worth making this function more automatic, but it should be one of the first places to check if there is a problem in the future with new running in new Jupyter environments. Check with Yan to confirm that this solves #67.