Option to ignore unattached nodes
When initializing D3ARG object from a tskit.TreeSequence, I've added an option to ignore nodes that are completely unattached (not in the tskit.TreeSequence edge table). This was mentioned in issue #69 to reduce the number of nodes displayed to the screen after simplifying with `filter_nodes=False`. I debated keeping the unattached nodes in the D3ARG object and adding the ability to ignore when drawing, but the current implementation seemed like a cleaner solution.