Figure titles and tree highlighting in node view
Addresses Issues #72 and #85. Titles do not allow for any styling changes. They also don't include wrapping so content should be kept brief.

Changed how tree highlighting works so that genome blocks can interactive or not. This allows for tree highlighting in the node view where certain blocks are not included in the subgraph. Those blocks are inactivated and ignored in highlighting.