Node view - ignoring unimportant breakpoints
When viewing a subgraph, not all of the breakpoints along the chromosome will be relevant. This addition combines tree highlighting regions when the breakpoint is not included within the subgraph. Tree highlighting works as before. I did not consider perform when coding this, so it may be inefficient for large subgraphs, though generally I don't think you would be plotting those with this method.