create a mutations branch
Here's a branch for addressing #14. I've implemented a method for adding mutations to the graph and merged most of the functionality of `draw()` and `draw_node()` so that updates address both. Previously, these were completely separate functions which made development hard. I plan to add parameters regarding how the mutations are included, as currently they are always added when `edge_type="line"` which isn't always ideal. Mutations do not work for `edge_type="ortho"` and at the moment that is a bigger undertaking than I can tackle.