fixed when no mutations and added genome bar standalone
Previously, when there weren't mutations, it had self.mutations as an empty pd.DataFrame. This is fine, but it's important that it has the same column names even if they are empty.

Changed edge "color" to edge "stroke" to be consistent with the other tables.

Also, working on a genome bar standalone plot that includes mutations and the ability to draw a windows. Made it so that genome blocks have a fill color that can be changed. We may want to eventually merge this with the visualizer where the visualizer calls the genome bar script rather than maintaining them separately.