change mutation count to be calculated on draw
Addressing Yan's issue in #14. I told him to comment out the mutation count, but it turns out mutation count is used to determine the positioning of the mutation marks. This is likely the issue with his plot. Mutation count is now being calculated at draw, which should also be much faster for large ARGs as only a subset of edges are included in the plot. I will need to come back and clean up the commented code after finalizing things.