mutation interactions
Regarding #14, I've start working on some of the interactions that we've discussed. Firstly, I've added two parameters `show_mutations` and `include_mutation_labels` to `draw()` and `draw_node()`. `show_mutations` just tells whether to include mutations in the plot at all; if False, an empty list is sent. `include_mutation_labels` determines whether to include the label within the mutation mark. By default, this is false which means that marks are small without a label - I prefer this  styling as I think it's the least cluttered. When you hover over a mutation mark, it shows where that mutation occurs along the genome bar along with a label (position:ancestral:derived). If you click the mark, it will lock it on until you click it again. "Active" mutations are shown with red for some visual feedback. This allows you to highlight a subset of mutations if you want to take a picture. Technically, you could preset mutations as "active" from the pandas.DataFrame, but I have not added a function to do this. If `include_mutation_labels=True`, the labels will be added to the mutations and on hover the text label on the genome bar is then (site_id:position) as the alleles are already displayed in the label.

Lastly before I forget, the positioning of mutations only works for `edge_type="line"`, though it still will fail to position properly if there are small diamonds as these cause the edges to curve.