#89 save as HTML
This turned out to be harder than I think it should be and might indicate a need to rethink things going forward. Not much code required but it seems hacky. Short walkthrough of the challenges - If we save the source file, it won't remember the new node positions. If we save the outerHTML object, that will lead to the original diagram and a new diagram being created because the JavaScript will run again. So I save the outerHTML and remove the div containing the old SVG. Now we are essentially back to the source file, but we still have the D3 force-layout in memory with all of the new node positions. So I find and replace all of the old node info with the new node info in the outerHTML string and save as a Blob. It doesn't feel the cleanest and I wish that there were a way to do this more directly with D3.js but I haven't found one.