Tooltip
#50 asks for popup info on hover. Had some time to implement this. Used a strategy similar to https://james-kitchens.com/blog/visualizing-human-genetic-diversity, where the tooltip is an extra div appended to the div containing the SVG. When you hover over an object that has popup information, it repositions the tooltip div to be below your mouse and fills the div with relevant information. This way it's always only one div rather than lots of popup divs that are being hidden or shown on hover. I added this for the mutations #14 as this is an instance where there is not enough room in the SVG to show all of the mutation information. Because of this, I removed the label from the genome bar mutation mark since I am now presenting that in the graph.

An additional thought is that this tooltip/extra info div could be its own subplot that can be turned on or off, similar to the genome bar. I'm starting to like the idea of many small subplots that communicate with one another rather than one plot that gets more and more complex. With this strategy, I think the tooltip would be fixed to a position or corner on the screen and be populated with info based on what you hover over. That seems like it could be cleaner though may not be as initially intuitive... I'm uncertain.