draw_node(..., return_included_nodes) and tutorial
GitHub Issues: #87, #94, #97, #98

Added a simple parameter to draw_node() to return the nodes that are included in the subgraph that is plotted. The other issues refer to needs for documentation. I've started a tutorial document to walk through the different plots and functions. Some is repeated in the README or other documentation, but this should be the starting place for people using the package. It will need to be updated as features are added or changes. I generally tried to avoid things that are less stable, such as how the package interacts with different environments since this is an ongoing issue.