#106 - rotating tip labels
Quick implementation of this. Required adding a group around the labels so that the rotation can be applied to the text without affecting the positioning. The implementation with `draw_node()` may not work as expected as only nodes without children will have rotated labels, even if those children are not drawn within the subgraph (thinking of @hyanwong's implementation here).