start of progressive ARGs and #108 hotfix
This adds the `zoom` parameter to `draw()` for collapsing graphs using a `progressive meshes` type technique. Additionally, there is a lot of bloat in the JSON from the `evenly_distributed_positions` that is not even usable with `draw_node()` so removing that for the time being until we come up with something better.

Other related GitHub issues - #93, #109