#125 - fix for some use cases
This issue was caused by "position_01" (the pre-scaled position along the chromosome of a mutation) not being passed to the JSON as it is not needed for plotting. It is needed if you download the JSON and load it into a D3ARG as otherwise you wouldn't have the ability to scale with plot sizes appropriately. What we really need is all of the information in the tables to be passed to the JSON rather than any sort of subset. This would ensure that as we add more features, downloading the JSON still works properly.

This commit does not fix what to do if you download the JSON from a plot with condensed mutations. The condensed mutations really screw things up and I haven't thought of a great way around it.