Updated label switching and new line label positioning
This is looking close. Because we are providing controls to the user within JS, I've removed the `include_labels` field from `D3ARG.nodes`. I updated the styling of the node labels options dropdown to match the others in the dashboard.

Previously, each node had multiple labels created as `<tspan>` elements with all but one hidden. This loads incorrectly in Adobe Illustrator (#127) as Illustrator does not respect `display:none` styling. To get around this, and generally be cleaner, I've made it so that changing the node label actually redraws the object with new text rather than hiding and showing things. This should scale better in the event we have more options.

Lastly, I positioned multiline labels so that the bottom of the label is always in the same position as the original numbers that we used to use. This may get confusing with really long labels as the center of the label could appear quite far above the actual node, but I don't think labels should necessary be long to begin with to avoid cluttering the screen.