solving JSON issues and removed commented `include_label`
The newline characters were leading to issues with the find and replace that is necessary for saving to JSON. Essentially, it was never able to replace the old source, which then led to missing columns. To solve this I escaped newline characters in the source before doing the find and replace. This still feels very hacky and fraught with issues and would greatly benefit from building modular functions in JS rather than using the template method that is currently implemented.

I may want to update the `num_samples` and `sample_order` attributes of the `D3ARG` as they feel somewhat out of place. Will think on this!