Customizable y-axis
Users can now set there own y-axis ticks using a list or dictionary. `_prepare_json()` has kind of gotten out of hand and could use some streamlining. Made it so that positions are all calculated at the start and then applied to the nodes, mutations, and labels. These functions could further be isolated by separating into functions. Checked that it works for a single node (even if that node is aat time 0). If `y_axis_labels` is a list or dictionary, default labels are ignored.