default node styling
You can now set a default node styling which is applied on `reset_node_styles()` rather than using the tskit colors that I've been defaulting to. These styling options are now passed to the JSON for plotting. They are not used during plotting but make it so that these options are set when loading a D3ARG from a downloaded JSON.