Fix downloading as SVG/PNG
This was a very difficult bug to track down, but the missing styles after download was due to including the class `d3arg` in the CSS rules. This class is necessary to avoid CSS rules spilling into unexpected places, but when you download the SVG it does not include the `d3arg` div so the styles aren't applied. I removed this class from the styles during download.

I've also updated the README figure and included the code used to generate it.