Adding escaped newline character and fixing #191
Found a bug when using condensed mutations where they weren't showing up on the genome bar. This is because not all of the necessary columns from d3arg.mutations was being passed along when condensing. Also found that newline characters need to be escaped so added a small replacement in the title as I think people will be more accustomed to not escaping.